Extract extractPlaceholderValuesFromFilePath into the utilities module
